Move validation actions to checkpoint module
