+++ great_expectations/checkpoint/__init__.py
@@ -1,1 +1,21 @@
+from .actions import (
+    MicrosoftTeamsNotificationAction,
+    NoOpAction,
+    OpsgenieAlertAction,
+    PagerdutyAlertAction,
+    SlackNotificationAction,
+    StoreEvaluationParametersAction,
+    StoreMetricsAction,
+    StoreValidationResultAction,
+    UpdateDataDocsAction,
+    ValidationAction,
+)
 from .checkpoint import Checkpoint, LegacyCheckpoint
+from ..util import verify_dynamic_loading_support
+
+for module_name, package_name in [
+    (".actions", "great_expectations.checkpoint"),
+    (".checkpoint", "great_expectations.checkpoint"),
+    (".util", "great_expectations.checkpoint")
+]:
+    verify_dynamic_loading_support(module_name=module_name, package_name=package_name)
