+++ tests/actions/test_store_metric_action.py
@@ -11,7 +11,7 @@ from great_expectations.data_context.types.resource_identifiers import (
     ExpectationSuiteIdentifier,
     ValidationResultIdentifier,
 )
-from great_expectations.validation_operators.actions import StoreMetricsAction
+from great_expectations.checkpoint.actions import StoreMetricsAction
 
 
 @freeze_time("09/26/2019 13:42:41")
